requirements.txt
psycopg2-binary
flask
requests
waitress
